Moved styled_input_button file
